move charset tokenizer to optional module
