+++ Cargo.toml
@@ -6,12 +6,16 @@ authors = ["emily Crandall Fleischman <emilycf@mit.edu>"]
 license = "MIT"
 
 [dependencies]
-bstr = "1.9.1"
-bytes = "1.6.0"
-itertools = "0.13"
+bstr = { version = "1.9.1", optional = true }
+bytes = { version = "1.6.0", optional = true }
+either = { version = "1.15.0", optional = true }
 thiserror = "1.0.38"
-unicode-xid = "0.2.4"
+unicode-xid = { version = "0.2.4", optional = true }
+
+[features]
+charset-tokenizer = ["dep:bstr", "dep:bytes", "dep:either", "dep:unicode-xid"]
 
 [dev-dependencies]
 anyhow = "1.0.68"
 test-case = "2.2.2"
+expr-parser = { path = ".", features = ["charset-tokenizer"] }
